+++ src/app/app.component.html
@@ -1,3 +1,1 @@
-<app-topbar />
-<app-main />
 <router-outlet />
