+++ src/app/app.component.ts
@@ -1,12 +1,11 @@
 import { Component } from '@angular/core';
 import { Route,RouterOutlet } from '@angular/router';
-import { TopbarComponent } from './topbar/topbar.component';
-import { MainComponent } from './main/main.component';
+import { TopbarComponent } from './topbar-Home/topbar.component';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet,TopbarComponent,MainComponent],
+  imports: [RouterOutlet,TopbarComponent],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss'
 })
