+++ src/app/app.routes.ts
@@ -1,6 +1,8 @@
 import { Routes } from '@angular/router';
-import { LoginComponent } from './login/login.component';
+import { LoginComponent } from './login-register/login.component';
+import { TopbarComponent } from './topbar-Home/topbar.component';
 
 export const routes: Routes = [
-    { path: 'login', component: LoginComponent}
+    { path: 'login', component: LoginComponent},
+    { path: '', component: TopbarComponent}
 ];
