+++ src/app/login-register/login.component.html
@@ -1,0 +1,66 @@
+<body>
+    <div class="container" [ngClass]="{'active': isBoolean}">
+        <div class="form-box login">
+            <form action="">
+                <h1>登入會員</h1>
+                <div class="input-box">
+                    <input type="text" placeholder="Username">
+                    <i class='bx bxs-user'></i>
+                </div>
+                <div class="input-box">
+                    <input type="password" placeholder="Password">
+                    <i class='bx bxs-lock-alt'></i>
+                </div>
+                <div class="forget-link">
+                    <a href="#">忘記密碼?</a>
+                </div>
+                <button type="submit" class="btn">登入</button>
+                <div class="social-icons">
+                    <a href="#"><i class='bx bxl-facebook-circle'></i></a>
+                    <a href="#"><i class='bx bxl-google'></i></a>
+                    <a href="#"><i class='bx bxl-github'></i></a>
+                </div>
+            </form>
+        </div>
+
+        <div class="form-box register">
+            <form action="">
+                <h1>註冊會員</h1>
+                <div class="input-box">
+                    <input type="text" placeholder="Username">
+                    <i class='bx bxs-user'></i>
+                </div>
+                <div class="input-box">
+                    <input type="email" placeholder="Email">
+                    <i class='bx bxs-envelope'></i>
+                </div>
+                <div class="input-box">
+                    <input type="password" placeholder="Password">
+                    <i class='bx bxs-lock-alt'></i>
+                </div>
+                <button type="submit" class="btn">註冊</button>
+                <div class="social-icons">
+                    <a href="#"><i class='bx bxl-facebook-circle'></i></a>
+                    <a href="#"><i class='bx bxl-google'></i></a>
+                    <a href="#"><i class='bx bxl-github'></i></a>
+                </div>
+            </form>
+        </div>
+
+        <div class="toggle-box">
+            <div class="toggle-panel toggle-left">
+                <h1>Hello, Welcome!</h1>
+                <p>Don't have an account?</p>
+                <button (click)="isBoolean = !isBoolean" class="btn register-btn">註冊會員</button>
+            </div>
+
+            <div class="toggle-panel toggle-right">
+                <h1>Welcome Back!</h1>
+                <p>Already have an account?</p>
+                <button (click)="isBoolean = !isBoolean" class="btn login-btn">登入會員</button>
+            </div>
+        </div>
+
+
+    </div>
+</body>
