+++ src/app/login-register/login.component.scss
@@ -19,6 +19,8 @@ body {
     background: #fff;
     border-radius: 30px;
     box-shadow: 0 0 30px rgb(0, 0, 0, 0.2);
+    overflow: hidden;
+    margin: 20px;
 }
 
 .form-box {
@@ -26,12 +28,27 @@ body {
     right: 0;
     width: 50%;
     height: 100%;
-    background: seagreen;
+    background: #fff;
     display: flex;
     align-items: center;
     color: #333;
     text-align: center;
     padding: 40px;
+    z-index: 1;
+    transition: 0.6s ease-in-out 1.2s, visibility 0s 1s;
+}
+
+.container.active .form-box {
+    right: 50%;
+}
+
+.form-box.register {
+    visibility: hidden;
+}
+
+.container.active .form-box.register {
+    visibility: visible;
+
 }
 
 form {
@@ -98,6 +115,11 @@ form {
     margin-bottom: 50px;
 }
 
+.social-icons {
+    display: flex;
+    justify-content: center;
+}
+
 .social-icons a {
     display: inline-flex;
     padding: 10px;
@@ -107,4 +129,71 @@ form {
     color: #333;
     text-decoration: none;
     margin: 0 10px;
+}
+
+.toggle-box {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+}
+
+.toggle-box::before {
+    content: '';
+    position: absolute;
+    left: -250%;
+    width: 300%;
+    height: 100%;
+    background: #7494ec;
+    border-radius: 150px;
+    z-index: 2;
+    transition: 1.8s ease-in-out;
+}
+
+.container.active .toggle-box::before {
+    left: 50%;
+}
+
+.toggle-panel {
+    position: absolute;
+    width: 50%;
+    height: 100%;
+    color: #fff;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    z-index: 2;
+    transition: 0.6s ease-in-out;
+}
+
+.toggle-panel.toggle-left {
+    left: 0;
+    transition-delay: 1.2s;
+}
+
+.container.active .toggle-panel.toggle-left {
+    left: -50%;
+    transition-delay: 0.6s;
+}
+
+.toggle-panel.toggle-right {
+    right: -50%;
+    transition-delay: 0.6s;
+}
+
+.container.active .toggle-panel.toggle-right {
+    right: 0;
+    transition-delay: 1.2s;
+}
+
+.toggle-panel p {
+    margin-bottom: 20px;
+}
+
+.toggle-panel .btn {
+    width: 160px;
+    height: 46px;
+    background: transparent;
+    border: 2px solid #fff;
+    box-shadow: none;
 }
