--- src/app/login/login.component.html
@@ -1,26 +1,0 @@
-<body>
-    <div class="container">
-        <div class="form-box">
-            <form action="">
-                <h1>登入會員</h1>
-                <div class="input-box">
-                    <input type="text" placeholder="Username">
-                    <i class='bx bxs-user'></i>
-                </div>
-                <div class="input-box">
-                    <input type="password" placeholder="Password">
-                    <i class='bx bxs-lock-alt'></i>
-                </div>
-                <div class="forget-link">
-                    <a href="#">忘記密碼?</a>
-                </div>
-                <button type="submit" class="btn">登入</button>
-                <div class="social-icons">
-                    <a href="#"><i class='bx bxl-facebook-circle'></i></a>
-                    <a href="#"><i class='bx bxl-google'></i></a>
-                    <a href="#"><i class='bx bxl-github'></i></a>
-                </div>
-            </form>
-        </div>
-    </div>
-</body>
