+++ src/app/main/main.component.scss
@@ -8,7 +8,7 @@
     padding-right: 3vw;
 }
 .survey-card-container {
-    height: 100vh;
+    height: 65vh;
     position: relative;
 }
 .survey-card {
@@ -22,37 +22,37 @@
 .survey-card-container .survey-card:hover {
     box-shadow: 0px 0px 7px rgb(0, 0, 0);
     filter: blur(0px);
-    transform: scale(1.5);
+    transform: scale(1.3);
     z-index: 5;
     transition: 0.5s;
 }
 
 .survey-card:nth-child(1){
-    background-color: rgba(14, 145, 174, 0.81);
+    background-color: rgba(14, 145, 174, 0.75);
     bottom: 24%;
     left: 28%;
     transform: rotate(15deg)
 }
 .survey-card:nth-child(2){
-    background-color: rgba(54, 193, 8, 0.8);
+    background-color: rgba(54, 193, 8, 0.75);
     bottom: 30%;
     left: 37%;
     transform: rotate(-13deg)
 }
 .survey-card:nth-child(3){
-    background-color: rgba(177, 10, 183, 0.8);
+    background-color: rgba(177, 10, 183, 0.75);
     bottom: 25%;
     left: 47%;
     transform: rotate(5deg)
 }
 .survey-card:nth-child(4){
-    background-color: rgba(240, 217, 6, 0.8);
+    background-color: rgba(240, 217, 6, 0.75);
     bottom: 30%;
     left: 57%;
     transform: rotate(-20deg)
 }
 .survey-card:nth-child(5){
-    background-color: rgba(165, 165, 161, 0.8);
+    background-color: rgba(239, 149, 30, 0.75);
     bottom: 22%;
     left: 66%;
     transform: rotate(26deg)
