+++ src/app/topbar-Home/topbar.component.scss
@@ -1,0 +1,75 @@
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+}
+
+body {
+    overflow: hidden;
+}
+
+.topbar {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    padding: 35px;
+}
+
+i {
+    font-size: 28px;
+    color: #757575;
+}
+
+.search {
+    position: relative;
+    z-index: 2;
+}
+
+
+.search-box.active {
+    position: absolute;
+    transform: translateX(300px);
+    background: rgba(164, 66, 249, 0.4);
+    z-index: 2;
+    transition: 0.7s;
+}
+
+.search-box.active .bx-left-arrow-alt {
+    position: absolute;
+    top: 0;
+    display: unset;
+    transform: translate(28px,5px);
+    color: rgb(157, 3, 3);
+}
+
+.bx-left-arrow-alt {
+    display: none;
+}
+
+.search-line {
+    position: absolute;
+    border-radius: 0px 28px 28px 0px;
+    top: 0;
+    width: 10px;
+    height: 40px;
+    z-index: -1;
+    transition: 0.5s ease-in-out;
+}
+
+.search-line.active {
+    transition: 0.5s ease-in-out;
+    width: 350px;
+    background: rgba(233, 115, 115, 0.75);
+}
+
+.search-line {
+    transition: 0.5s;
+    width: 10px;
+}
+
+.search p,.member p {
+    font-size: 10px;
+    font-weight: 500;
+    color: #757575;
+}
+
