--- src/app/topbar/topbar.component.html
@@ -1,1 +1,0 @@
-<p>topbar works!</p>
