--- src/app/topbar/topbar.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-topbar',
-  standalone: true,
-  imports: [],
-  templateUrl: './topbar.component.html',
-  styleUrl: './topbar.component.scss'
-})
-export class TopbarComponent {
-
-}
